Update project according to pull request comments
