+++ homework-3-csv-app/src/main/java/org/homework/config/CsvReaderConfig.java
@@ -2,6 +2,7 @@ package org.homework.config;
 
 import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
+import org.homework.model.Exam;
 import org.homework.utils.reader.CsvReader;
 import org.homework.utils.reader.CsvReaderImpl;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,15 +17,28 @@ import java.io.FileReader;
 @Configuration
 public class CsvReaderConfig {
     private final File file;
+    private final Integer passBorder;
 
-    public CsvReaderConfig(@Value("${application.fileName}_${application.locale}.csv") String fileName) throws FileNotFoundException {
+    public CsvReaderConfig(@Value("${application.fileName}_${application.locale}.csv") String fileName, @Value("${application.passBorder}") Integer passBorder) throws FileNotFoundException {
         file = ResourceUtils.getFile("classpath:" + fileName);
+        checkBorderValue(passBorder);
+        this.passBorder = passBorder;
     }
 
-    @Bean
     public CsvReader csvReader() throws FileNotFoundException {
         FileReader fileReader = new FileReader(file);
         CSVReader csvReader = new CSVReaderBuilder(fileReader).withSkipLines(1).build();
         return new CsvReaderImpl(csvReader, fileReader);
+    }
+
+    @Bean
+    public Exam exam() throws Exception {
+        return csvReader().getAsExam(passBorder);
+    }
+
+    private void checkBorderValue(Integer passBorder) {
+        if (!((passBorder >= 0) && (passBorder <= 100))) {
+            throw new RuntimeException("Incorrect border value. Value should be: 0 >= value <= 100");
+        }
     }
 }
