+++ homework-3-csv-app/src/main/resources/i18n/bundle_ru_RU.properties
@@ -1,12 +1,12 @@
-msg.start.welcome=Добро пожаловать на базовый экзамен по java!
-msg.start.firstName=Пожалуйста, введите имя
-msg.start.lastName=Пожалуйста, введите фамилию
-msg.start.greeting=Привет
-msg.inprogress.answer.text=Ваш ответ
-msg.inprogress.answer.negativeresult=не правильный
-msg.inprogress.answer.positiveresult=правильный
-msg.inprogress.answer.empty=Пустой ответ
-msg.finish.result=Результат вашего экзамена
-msg.finish.result.border=Минимальный результат для сдачи
-msg.finish.result.passed=Поздравляем! Вы сдали экзамен!
-msg.finish.result.failed=Вы не сдали экзамен. Попробуйте позже
+msg.start.welcome=Р”РѕР±СЂРѕ РїРѕР¶Р°Р»РѕРІР°С‚СЊ РЅР° Р±Р°Р·РѕРІС‹Р№ СЌРєР·Р°РјРµРЅ РїРѕ java!
+msg.start.firstName=РџРѕР¶Р°Р»СѓР№СЃС‚Р°, РІРІРµРґРёС‚Рµ РёРјСЏ
+msg.start.lastName=РџРѕР¶Р°Р»СѓР№СЃС‚Р°, РІРІРµРґРёС‚Рµ С„Р°РјРёР»РёСЋ
+msg.start.greeting=РџСЂРёРІРµС‚
+msg.inprogress.answer.text=Р’Р°С€ РѕС‚РІРµС‚
+msg.inprogress.answer.negativeresult=РЅРµ РїСЂР°РІРёР»СЊРЅС‹Р№
+msg.inprogress.answer.positiveresult=РїСЂР°РІРёР»СЊРЅС‹Р№
+msg.inprogress.answer.empty=РџСѓСЃС‚РѕР№ РѕС‚РІРµС‚
+msg.finish.result=Р РµР·СѓР»СЊС‚Р°С‚ РІР°С€РµРіРѕ СЌРєР·Р°РјРµРЅР°
+msg.finish.result.border=РњРёРЅРёРјР°Р»СЊРЅС‹Р№ СЂРµР·СѓР»СЊС‚Р°С‚ РґР»СЏ СЃРґР°С‡Рё
+msg.finish.result.passed=РџРѕР·РґСЂР°РІР»СЏРµРј! Р’С‹ СЃРґР°Р»Рё СЌРєР·Р°РјРµРЅ!
+msg.finish.result.failed=Р’С‹ РЅРµ СЃРґР°Р»Рё СЌРєР·Р°РјРµРЅ. РџРѕРїСЂРѕР±СѓР№С‚Рµ РїРѕР·Р¶Рµ
